requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
SQLAlchemy
passlib
bcrypt
python-jose
cryptography
python-multipart
python-decouple
psycopg2
alembic
fastapi-cors
fastapi-mail




